Adicionado endpoints, criar estudar, criar turma e criar professor
