+++ src/endpoints/criarEstudante.ts
@@ -1,19 +1,22 @@
-import app from "../app";
 import {Request, Response} from "express";
-import {connection} from "../connection";
+import adicionarEstudante from "../data/adicionarEstudante";
 
-app.put("/estudante/criar", async (req: Request, res: Response) => {
+export default async function criarEstudante(
+    req: Request, res: Response
+) {
     try {
-        await connection.raw(`
-            INSERT INTO Student (nome, email, data_nasc, hobbies)
-            VALUES ("${req.body.nome}",
-                    "${req.body.email}",
-                    "${req.body.data_nasc}",
-                    "${req.body.hobbies}")
-        `)
-        res.status(200).send("Estudante criado com sucesso!")
+        await adicionarEstudante(
+            req.body.nome,
+            req.body.email,
+            req.body.data_nasc,
+            req.body.turma_id
+        )
 
+        res.status(200)
+            .send("Estudante criado com sucesso!")
     } catch (error) {
-        res.status(400).send("Verifique se os campos foram preenchidos corretamente.")
+        res
+            .status(400)
+            .send({message: error.message || error.sqlMessage})
     }
-})
+}
