+++ src/endpoints/criarTurma.ts
@@ -1,20 +1,31 @@
-import app from "../app";
-import {Request, Response} from "express";
-import {connection} from "../connection";
+import {Request, Response} from "express"
+import adicionarTurma from "../data/adicionarTurma";
 
-app.put("/turma/criar", async (req: Request, res: Response) => {
+export default async function criarTurma(
+    req: Request, res: Response
+) {
     try {
-        await connection.raw(`
-            INSERT INTO Class (nome, data_de_inicio, data_de_encerramento, professores, modulo)
-            VALUES ("${req.body.nome}",
-                    "${req.body.data_de_inicio}",
-                    "${req.body.data_de_encerramento}",
-                    "${req.body.professores}",
-                    "${req.body.modulo}")
-        `)
-        res.status(200).send("Turma criada com sucesso!")
+        if (!req.body.nome || !req.body.data_inicio || !req.body.data_final || !req.body.modulo) {
+            throw new Error("Por favor, preencha todos os dados.")
+        }
 
+        if (req.body.modulo <= 7) {
+            throw new Error("Por favor, verifique modulo.")
+        }
+
+        await adicionarTurma(
+            req.body.nome,
+            req.body.data_inicio,
+            req.body.data_final,
+            req.body.modulo
+        )
+
+        res
+            .status(200)
+            .send("Turma criada com sucesso !")
     } catch (error) {
-        res.status(400).send("Verifique se os campos foram preenchidos corretamente.")
+        res
+            .status(400)
+            .send({message: error.message || error.sqlMessage})
     }
-})
+}
