--- src/endpoints/criarTurma2.ts
@@ -1,27 +1,0 @@
-import {Request, Response} from "express"
-import adicionarTurma from "../data/adicionarTurma";
-
-export default async function criarTurma(
-    req: Request, res: Response
-) {
-    try {
-        if (!req.body.nome || !req.body.data_inicio || !req.body.data_final || !req.body.modulo) {
-            throw new Error("Por favor, preencha todos os dados.")
-        }
-
-        await adicionarTurma(
-            req.body.nome,
-            req.body.data_inicio,
-            req.body.data_final,
-            req.body.modulo
-        )
-
-        res
-            .status(200)
-            .send("Turma criada com sucesso !")
-    } catch (error) {
-        res
-            .status(400)
-            .send({message: error.message || error.sqlMessage})
-    }
-}
